Move frame type to common
